+++ src/pudl/etl.py
@@ -20,7 +20,7 @@ import time
 from concurrent.futures import ProcessPoolExecutor
 from functools import partial
 from pathlib import Path
-from typing import Any
+from typing import Any, Literal
 
 import pandas as pd
 import pyarrow as pa
@@ -29,8 +29,7 @@ import sqlalchemy as sa
 
 import pudl
 from pudl.helpers import convert_cols_dtypes
-from pudl.metadata.classes import Resource
-from pudl.metadata.codes import CODE_METADATA
+from pudl.metadata.classes import Package, Resource
 from pudl.metadata.dfs import FERC_ACCOUNTS, FERC_DEPRECIATION_LINES
 from pudl.metadata.fields import apply_pudl_dtypes
 from pudl.settings import (
@@ -49,30 +48,6 @@ logger = logging.getLogger(__name__)
 ###############################################################################
 # EIA EXPORT FUNCTIONS
 ###############################################################################
-
-
-def _read_static_tables_eia() -> dict[str, pd.DataFrame]:
-    """Build dataframes of static EIA tables for use as foreign key constraints.
-
-    There are many values specified within the data that are essentially
-    constant, but which we need to store for data validation purposes, for use
-    as foreign keys.  E.g. the list of valid EIA fuel type codes, or the
-    possible state and country codes indicating a coal delivery's location of
-    origin.
-
-    """
-    return {
-        "energy_sources_eia": CODE_METADATA["energy_sources_eia"]["df"],
-        "operational_status_eia": CODE_METADATA["operational_status_eia"]["df"],
-        "fuel_types_aer_eia": CODE_METADATA["fuel_types_aer_eia"]["df"],
-        "prime_movers_eia": CODE_METADATA["prime_movers_eia"]["df"],
-        "sector_consolidated_eia": CODE_METADATA["sector_consolidated_eia"]["df"],
-        "fuel_transportation_modes_eia": CODE_METADATA["fuel_transportation_modes_eia"][
-            "df"
-        ],
-        "contract_types_eia": CODE_METADATA["contract_types_eia"]["df"],
-        "coalmine_types_eia": CODE_METADATA["coalmine_types_eia"]["df"],
-    }
 
 
 def _etl_eia(
@@ -102,7 +77,7 @@ def _etl_eia(
         return []
 
     # generate dataframes for the static EIA tables
-    out_dfs = _read_static_tables_eia()
+    out_dfs = _read_static_encoding_tables(etl_group="static_eia")
 
     ds = Datastore(**ds_kwargs)
     # Extract EIA forms 923, 860
@@ -148,9 +123,7 @@ def _etl_eia(
 
     for table in entities_dfs:
         entities_dfs[table] = (
-            pudl.metadata.classes.Package.from_resource_ids()
-            .get_resource(table)
-            .encode(entities_dfs[table])
+            Package.from_resource_ids().get_resource(table).encode(entities_dfs[table])
         )
 
     out_dfs.update(entities_dfs)
@@ -163,25 +136,25 @@ def _etl_eia(
 ###############################################################################
 
 
-def _read_static_tables_ferc1() -> dict[str, pd.DataFrame]:
-    """Populate static PUDL tables with constants for use as foreign keys.
+def _read_static_tables_ferc1():
+    """Compile static tables for FERC1 for foriegn key constaints.
 
-    There are many values specified within the data that are essentially
-    constant, but which we need to store for data validation purposes, for use
-    as foreign keys.  E.g. the list of valid EIA fuel type codes, or the
-    possible state and country codes indicating a coal delivery's location of
-    origin. For now these values are primarily stored in a large collection of
-    lists, dictionaries, and dataframes which are specified in the
-    pudl.metadata module.  This function uses those data structures to
-    populate a bunch of small infrastructural tables within the PUDL DB.
+    This function grabs static encoded tables via :func:`_read_static_encoding_tables`
+    as well as two static tables that are non-encoded tables (``ferc_accounts`` and
+    ``ferc_depreciation_lines``).
     """
-    return {
-        "ferc_accounts": FERC_ACCOUNTS[["ferc_account_id", "ferc_account_description"]],
-        "ferc_depreciation_lines": FERC_DEPRECIATION_LINES[
-            ["line_id", "ferc_account_description"]
-        ],
-        "power_purchase_types_ferc1": CODE_METADATA["power_purchase_types_ferc1"]["df"],
-    }
+    static_table_dict = _read_static_encoding_tables("static_ferc1")
+    static_table_dict.update(
+        {
+            "ferc_accounts": FERC_ACCOUNTS[
+                ["ferc_account_id", "ferc_account_description"]
+            ],
+            "ferc_depreciation_lines": FERC_DEPRECIATION_LINES[
+                ["line_id", "ferc_account_description"]
+            ],
+        }
+    )
+    return static_table_dict
 
 
 def _etl_ferc1(
@@ -357,7 +330,7 @@ def etl_epacems(
 
 
 ###############################################################################
-# GLUE EXPORT FUNCTIONS
+# GLUE AND STATIC EXPORT FUNCTIONS
 ###############################################################################
 def _etl_glue(
     glue_settings: GlueSettings,
@@ -414,6 +387,34 @@ def _etl_glue(
         glue_dfs.update(glue_transformed_dfs)
 
     return glue_dfs
+
+
+def _read_static_encoding_tables(
+    etl_group: Literal["static_eia", "static_ferc1"]
+) -> dict[str, pd.DataFrame]:
+    """Build dataframes of static tables from a data source for use as foreign keys.
+
+    There are many values specified within the data that are essentially constant, but
+    which we need to store for data validation purposes, for use as foreign keys.  E.g.
+    the list of valid EIA fuel type codes, or the possible state and country codes
+    indicating a coal delivery's location of origin. For now these values are primarily
+    stored in a large collection of lists, dictionaries, and dataframes which are
+    specified in the :mod:`pudl.metadata` subpackage.  This function uses those data
+    structures to populate a bunch of small infrastructural tables within the PUDL DB.
+
+    Args:
+        etl_group: name of static table etl group.
+
+    Returns:
+        a dictionary with table names as keys and dataframes as values for all tables
+        labeled as static tables in their resource ``etl_group``
+
+    """
+    return {
+        r.name: r.encoder.df
+        for r in Package.from_resource_ids().resources
+        if r.etl_group == etl_group and r.encoder
+    }
 
 
 ###############################################################################