+++ src/pudl/extract/eia923.py
@@ -42,6 +42,14 @@ class Extractor(excel.GenericExtractor):
         df = df.rename(columns=self._metadata.get_column_map(page, **partition))
         self.cols_added = []
         df = remove_leading_zeros_from_numeric_strings(df, "generator_id")
+        # the 2021 early release data had some ding dang "."'s and nulls in the year column
+        if "report_year" in df.columns:
+            mask = (df.report_year == ".") | df.report_year.isnull()
+            logger.debug(
+                f"{page}: replacing {len(df[mask])} nulls/bad values in `report_year` column with {partition['year']}"
+            )
+            df.loc[mask, "report_year"] = partition["year"]
+        df = self.add_data_maturity(df, page, **partition)
         return df
 
     def extract(self, settings: Eia923Settings = Eia923Settings()):